ARM: fix compilation with C++11 enabled
C++11 makes narrowing initialisation using {}s an error.
For an ARM target qreal != double, so force-cast to
qreal.

KAPPA now is consistent with the define of the same name
in qbezier.cpp.

+++ src/gui/painting/qpainterpath_p.h
@@ -270,7 +270,7 @@ inline void QPainterPathData::maybeMoveTo()
     }
 }
 
-#define KAPPA 0.5522847498
+#define KAPPA qreal(0.5522847498)
 
 
 QT_END_NAMESPACE
